backend/requirements.txt
boto3==1.17.53
botocore==1.20.53
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-auth==1.29.0
idna==2.10
jmespath==0.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.7
six==1.15.0
urllib3==1.26.4
